core: new unit testing framework
To overcome some of the limitations of the old unit test approach of
mbeddr, that was primary based on simple generate code, a new framework
was written in C and it now instrumented by the unit test generator.
The new approach solves two major problem:

First: illegal memory access in the tested code would abort the complete
test sequence because all tests were executed in the same process

Second: no machine readable result

To overcome the first one we now execute each test in a child process
and check if it failed. This can also detect process termination due
to illegal memory access. In addition it provides better test isolation.
A previous test can now not manipulate the program state of later executed
tests. This might require some changes in your tests because some test
might rely on the old undocumented approach.

The second is addressed by generating a TestResult.xml file which follows
JUnit xml schema. Each test execution also generated a .asserts.xml file
next to the binary which logs the executions of all asserts with its
expected and actual values. In addition it provides information about
failed asserts in case there are any.

The xml generated for each assert looks like this:

<assert expected="10" actual="10" nodeId="5611082447683490264" />

If the assert failed an additional entry:

<assert-fail expected="true" actual="false" nodeId="5611082447686603501"> assert failed </assert-fail>

is generated. The content of this tag is the fail message. Which indicates
what kind of assertion has failed.