ccmenu: New language for customizing the code completion menu
You can change the order of entries, sort entries into tabs and change
the color of entries.

There is already a tab definied, that allows you to create a new
reference target (you can type the name of a variable, that does not
exist and create this variable and a reference to it at the same time).