TopDownLayout: New framework inspired by CSS and AWT layout managers
Layouting happens from the root to the leafs. That means, children are
asked for their preferred/min/max size, but the parent sets the size and
the child has to fit itself into that size.

This gives more control over the resulting layout, such as:
- Richtext (or any ident layout) can be wrapped to fit into the size of
  a table column. At the moment indent layout always wrap at the right
  margin line.
- Make a horizontal line a wide as the cell above/below or as the whole
  page.
- Have additional cells on the right of a richtext without them being
  push over the right margin (the text will wrap earlier).
- In a diagram, the content of a box can fit to the size of the box.
  This is especially useful is there are subdiagrams in the box. Before,
  this was possible only with some ugly hacks.

There is also a CSS like box model that allows to set border sizes and
margins on cells and respect them in the layout. MPS by default allows
only borders of size 1.

This framework is compatible to the layouts existing layouts in MPS.
That means, there are implementations for horizontal/vertical/
vertical-grid/indent layouts, that support this new architecture. You
don't to change anything and you can still use the MPS editor language.

You can also have custom layouts base on the default MPS layout
architecture. This framework is able to handle them in a parent cell
(e. g. margin comments) or in a child cell (e. g. math).