ProjectView: NPE in CustomProjectView.removeListeners
It seems that the IViewDescriptor is already unregistered at that time.
Listeners are now initialized right before they are registered. If one
of them is configured to be disabled, it will never be initialized.

When removing the listener, we can just check if it is != null and don't
need the descriptor anymore.

Fixes #1822