GrammarCells: Hints are now considered for enabling/disabling actions
Usually actions generated by grammar cells are disabled when the editor
is overridden. This didn't work when an editor was defined for the same
concept but with different hints.
Since actions are defined in the editor aspect we have access to the
EditorContext and can figure out the hints of the current cursor
position.