mpsutil: fix projection mode switcher gen
If a projection mode switcher with multiple options was used the emited
code wasn't correct. The generator used a `removeWhere` operation which
inconstrast to other operations on lists doesn't only add a filter but
mutates the list inplace. It was replaced with a normal `where` which
only acts as filter around the original list.