refactored access to LibraryInitializer in spawner
Because LibraryInitializer was retrieved using MPSCoreComponents.getLibraryInitializer(),
during a shutdown this led to an exception - MPSCoreComponents had been disposed earlier
and thus did not contain a reference to LibraryInitializer anymore. In a headless scenario
where MPS was started with a spawner plugin, this caused the MPS process to finish with
an error code.