doc, core, cc: remove breadcrumb from editor
The breadcrumb runtime is diabled for a long time, it makes no sense
to have the breadcrumb editor cells around and to generated code into
the editors for breadcrumb handling. It seems like this code is also
creating problems when documentation language is in presentation mode
since it tries to attach selection listeners twice.