Add dependency on tuples runtime to avoid compilation failure
A new interface Dictionary has been introduced to tuples runtime,
and must be on the classpath of this module since IResource extends Dictionary.
The dependency must be present in the "languages engaged in generation" list,
because now Java compiler will reuse the list of runtime solutions from the
cache created during generation.