build: fix mbeddr build
When mbeddr is built on CI the pre-built mbeddr platform is downloaded
to `artifacts`. This means that the `com.mbeddr.build` project sees
the platform build scripts twice, once from its own modules and once
from the artifacts. This confuses the build script generation.

The fix is to make sure that `pluginRoots` passed to the build script
generation do not contain the mbeddr platform artifact.

One way to achieve it would be to simply pass
`artifacts/de.itemis.mps.extensions` instead of `artifacts` but we can
also use this opportunity to separate dependencies (MPS-extensions) from
sibling project artifacts (mbeddr platform) and download dependencies to
`build/dependencies`.